# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface ContinueReading {
  studySetId: string;
  documentId: string;
  studySetTitle: string;
  documentName: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @minimum 1 */
  page: number;
}

export interface DashboardSummary {
  totalStudySets: number;
  totalDocuments: number;
  totalNotes: number;
  totalPoints: number;
  currentStreak: number;
  flaggedCount: number;
  continueReading: ContinueReading | null;
}

export interface StudySet {
  id: string;
  title: string;
  createdAt: string;
  updatedAt: string;
  documentCount: number;
  noteCount: number;
  bestScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @nullable */
  lastOpenedAt: string | null;
}

export type DocumentProcessingStatus = typeof DocumentProcessingStatus[keyof typeof DocumentProcessingStatus];


export const DocumentProcessingStatus = {
  uploaded: 'uploaded',
  processing: 'processing',
  ready: 'ready',
  needs_review: 'needs_review',
} as const;

export interface Document {
  id: string;
  name: string;
  size: number;
  objectPath: string;
  /** @nullable */
  pageCount: number | null;
  processingStatus: DocumentProcessingStatus;
  uploadedAt: string;
}

export type StudySetDetail = StudySet & {
  documents: Document[];
};

export interface DocumentInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  objectPath: string;
}

export interface StudySetInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /**
     * @minItems 1
     * @maxItems 5
     */
  documents: DocumentInput[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: UploadUrlRequest;
}