# lib/api-zod/src/generated/types/continueReading.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContinueReading {
  studySetId: string;
  documentId: string;
  studySetTitle: string;
  documentName: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @minimum 1 */
  page: number;
}