# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContinueReading } from './continueReading';

export interface DashboardSummary {
  totalStudySets: number;
  totalDocuments: number;
  totalNotes: number;
  totalPoints: number;
  currentStreak: number;
  flaggedCount: number;
  continueReading: ContinueReading | null;
}