# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentProcessingStatus } from './documentProcessingStatus';

export interface Document {
  id: string;
  name: string;
  size: number;
  objectPath: string;
  /** @nullable */
  pageCount: number | null;
  processingStatus: DocumentProcessingStatus;
  uploadedAt: Date;
}