# lib/api-zod/src/generated/types/studySet.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudySet {
  id: string;
  title: string;
  createdAt: Date;
  updatedAt: Date;
  documentCount: number;
  noteCount: number;
  bestScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @nullable */
  lastOpenedAt: Date | null;
}