# lib/api-zod/src/generated/types/studySetDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';
import type { StudySet } from './studySet';

export type StudySetDetail = StudySet & {
  documents: Document[];
};