# lib/api-zod/src/generated/types/studySetInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentInput } from './documentInput';

export interface StudySetInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /**
     * @minItems 1
     * @maxItems 5
     */
  documents: DocumentInput[];
}